+++ tests/testthat/test-merge.R
@@ -24,30 +24,28 @@ test_that("merge doesn't work when it shouldn't", {
 
 test_that("merge returns the correct result", {
   epr3 <- merge(epr1, epr2)
-  expect_equal(
-    as.matrix(epr3),
-    rbind(
-      matrix(1:8, 4, 2, dimnames = list(letters[1:4], 1:2)),
-      matrix(1:8, 4, 2, dimnames = list(letters[5:8], 1:2))
-    )
+  res <- rbind(
+    matrix(1:8, 4, 2, dimnames = list(letters[1:4], 1:2)),
+    matrix(1:8, 4, 2, dimnames = list(letters[5:8], 1:2))
   )
+  names(dimnames(res)) <- c("levels", "time")
+  expect_equal(as.matrix(epr3), res)
   expect_identical(levels(epr3), letters[1:8])
   expect_identical(time(epr3), time(epr1))
   expect_equal(
     contrib(epr3, "a"),
-    matrix(c(0, 4), 1, 2, dimnames = list("a.1", 1:2))
+    matrix(c(0, 4), 1, 2, dimnames = list(levels = "a.1", time = 1:2))
   )
   expect_equal(
     contrib(epr3, "e"),
-    matrix(numeric(0), 0, 2, dimnames = list(NULL, 1:2))
+    matrix(numeric(0), 0, 2, dimnames = list(levels = NULL, time = 1:2))
   )
-  expect_equal(
-    as.matrix(merge(epr2, epr1)),
-    rbind(
-      matrix(1:8, 4, 2, dimnames = list(letters[5:8], 1:2)),
-      matrix(1:8, 4, 2, dimnames = list(letters[1:4], 1:2))
-    )
+  res2 <- rbind(
+    matrix(1:8, 4, 2, dimnames = list(levels = letters[5:8], 1:2)),
+    matrix(1:8, 4, 2, dimnames = list(levels = letters[1:4], 1:2))
   )
+  names(dimnames(res2)) <- c("levels", "time")
+  expect_equal(as.matrix(merge(epr2, epr1)), res2)
 })
 
 test_that("merge works with [", {